+++ CHANGES.rst
@@ -10,5 +10,8 @@ Added:
 - Added configurable section colors to configlet. #5
   [netroxen]
 
+- Added Clean and Demo profiles with demo content. #6
+  [netroxen]
+
 - Initial release.
   [sarn0ld]